+++ src/components/chartsV2/BarChart/BaseVerticalBarChart.stories.mdx
@@ -6,9 +6,8 @@ import { ThemedContainer } from '../../helpers';
 import {
 	sampleData,
 	sampleStackData,
-	transformToStack,
 	stackData,
-} from './BaseVerticalBarChartData.js';
+} from './BaseVerticalBarChartData.ts';
 import BaseBarChart from './BaseVerticalBarChart';
 
 <Meta title='Components/ChartsV2/BarChart' parameters={{ options: { showToolbar: false } }} />
