+++ src/components/chartsV2/BarChart/Skeleton/index.ts
