+++ src/components/chartsV2/BarChart/Skeleton/types/index.ts
@@ -1,0 +1,4 @@
+export type ChartSkeletonProps = {
+  theme?: any;
+  fallback?: boolean;
+};
