+++ src/components/chartsV2/BarChart/index.ts
