+++ src/components/chartsV2/pieChart/BasePieChart.stories.mdx
@@ -4,7 +4,7 @@ import { COLORS } from '../../../styles';
 import { ThemedContainer } from '../../helpers';
 import BasePieChart from './BasePieChart';
 
-import { sampleData, iconData, iconColor, piechartRandomColors } from './BasePieChartData.js';
+import { sampleData, iconData, iconColor, piechartRandomColors } from './BasePieChartData.ts';
 
 <Meta title='Components/ChartsV2/PieChart' parameters={{ options: { showToolbar: false } }} />
